Update ABCD login/logout for OPAC directory structure
The OPAC application is since pull request  Opac v2 #574 the main application and ABCD can be found via /admin.
The effect that logout/change password/wrong credentials/wrong captcha returned to OPAC in stead of the ABCD login is in this commit corrected.
Additionally a "OPAC" button is added to de login window.