Improve display of footer + html code
In a previous commit the open form in scripts_dataentry.php was closed in that file.
This happened to be wrong.
This commit reverses scripts_dataentry.php to the status with an open form. The other files in this commit were modified to get the footer back in all situations. Several illegal html statements were also addressed.
This is not completely resolved but the system functions as expected during the tests